+++ mex/extractors/contact_point/main.py
@@ -1,10 +1,10 @@
 from dagster import asset
 
+from mex.common.cli import entrypoint
 from mex.common.models import ExtractedContactPoint, ExtractedPrimarySource
-from mex.common.types import Email
+from mex.extractors.pipeline import run_job_in_process
+from mex.extractors.settings import Settings
 from mex.extractors.sinks import load
-
-MEX_EMAIL = "mex@rki.de"
 
 
 @asset(group_name="contact_point")
@@ -12,9 +12,16 @@ def extracted_contact_point_mex(
     extracted_primary_source_mex: ExtractedPrimarySource,
 ) -> None:
     """Load and return blueant primary source."""
+    settings = Settings.get()
     extracted_contact_point = ExtractedContactPoint(
-        identifierInPrimarySource=MEX_EMAIL,
+        identifierInPrimarySource=str(settings.contact_point.mex_email),
         hadPrimarySource=extracted_primary_source_mex.stableTargetId,
-        email=[Email(MEX_EMAIL)],
+        email=[settings.contact_point.mex_email],
     )
     load([extracted_contact_point])
+
+
+@entrypoint(Settings)
+def run() -> None:  # pragma: no cover
+    """Run the datscha-web extractor job in-process."""
+    run_job_in_process("contact_point")