+++ CHANGELOG.md
@@ -11,11 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - enable mypy linting for tests folder
 - include persons from allowed primary sources in the publishing pipeline
+- add contact-point mini extractor to get default contact points
 
 ### Changes
 
 - summarize logging for synopse filtering functions
-- filter out unpublished persons from contact fields and set mex@rki.de as fallback
+- filter out unpublished persons from contact fields and set a fallback
 
 ### Deprecated